+++ flh_opt/api_opt.py
@@ -2,10 +2,13 @@
 """API interface for FLH optimizer."""
 import math
 import os
-from typing import List, Optional
+from typing import List, Literal, Optional
 
 import pandas as pd
 from pypsa import Network
+from pypsa.descriptors import get_bounds_pu
+from pypsa.optimization.common import reindex
+from xarray import DataArray
 
 from flh_opt._types import OptInputDataType, OptOutputDataType
 
@@ -63,8 +66,8 @@ def _add_link(
             # input data is per main output,
             # pypsa link parameters are defined per main input
             capital_cost=(input_data[name]["CAPEX_A"] + input_data[name]["OPEX_F"])
-            / input_data[name]["EFF"],
-            marginal_cost=input_data[name]["OPEX_O"] / input_data[name]["EFF"],
+            * input_data[name]["EFF"],
+            marginal_cost=input_data[name]["OPEX_O"] * input_data[name]["EFF"],
             p_nom_extendable=True,
         )
         # add conversion efficiencies and buses for secondary input / output
@@ -73,8 +76,73 @@ def _add_link(
             # input data is per main output,
             # pypsa link parameters are defined per main input
             n.links.at[name, f"efficiency{i+2}"] = (
-                -input_data[name]["CONV"][c] / input_data[name]["EFF"]
+                -input_data[name]["CONV"][c] * input_data[name]["EFF"]
             )
+
+
+def get_flh(n: Network, g: str, component_type: Literal["Generator", "Link"]) -> float:
+    """Calculate full load hours.
+
+    Returns a value between 0 and 1.
+    """
+    if component_type == "Generator":
+        sw = n.snapshot_weightings["generators"]
+        gen = (n.generators_t["p"][g] * sw).sum()
+        p_nom = n.generators.at[g, "p_nom_opt"]
+    if component_type == "Link":
+        sw = n.snapshot_weightings["generators"]
+        gen = (n.links_t["p0"][g] * sw).sum()
+        p_nom = n.links.at[g, "p_nom_opt"]
+    if gen == 0:
+        flh = 0
+    else:
+        flh = gen / p_nom / 8760
+    return flh
+
+
+def scale_storage_soc_upper_bounds(n: Network):
+    """Scale the upper bounds of storage SOC with snapshot weightings.
+
+    We need to do this because of the week scaling and the fixed correlation
+    between charge capacity and state of charge for electricity storage.
+    In the storage balance, the effect of charging and discharging on SOC
+    is scaled with snapshot weightings.
+
+    This function also scales the storage capacity itself
+    with the snapshot weightings.
+    """
+    # if model has not yet been created, do it now:
+    if not hasattr(n, "model"):
+        n.optimize.create_model()
+
+    # get list of extendable storage units:
+    ext_i = n.get_extendable_i("StorageUnit")
+
+    # get max_hours attribute of these storage units:
+    max_hours = get_bounds_pu(
+        n, "StorageUnit", n.snapshots, index=ext_i, attr="state_of_charge"
+    )[1]
+
+    # multiply max_hours with snapshot weightings:
+    scaled_bounds = max_hours.copy()
+    for c in max_hours.columns:
+        scaled_bounds[c] = max_hours[c] * n.snapshot_weightings["stores"]
+    sb = DataArray(scaled_bounds, dims=["snapshot", "StorageUnit-ext"])
+
+    # get state of charge and charge capacity variables:
+    soc = reindex(
+        n.model.variables["StorageUnit-state_of_charge"], "StorageUnit", ext_i
+    )
+    p_nom = n.model.variables["StorageUnit-p_nom"]
+
+    # create left hand side of equation:
+    lhs = soc - p_nom * sb
+
+    # remove old constraint:
+    n.model.remove_constraints("StorageUnit-ext-state_of_charge-upper")
+
+    # and add the new one:
+    n.model.add_constraints(lhs, "<=", 0, name="StorageUnit-ext-state_of_charge-upper")
 
 
 def optimize(
@@ -404,28 +472,20 @@ def optimize(
 
     n.snapshot_weightings["generators"] = weights
     n.snapshot_weightings["objective"] = weights
-    n.snapshot_weightings["stores"] = 1
+    n.snapshot_weightings["stores"] = weights
 
     # import profiles to network:
     n.import_series_from_dataframe(res_profiles, "Generator", "p_max_pu")
 
+    # scale storage SOC constraints:
+    scale_storage_soc_upper_bounds(n)
+
     # solve optimization problem:
-    model_status = n.optimize(solver_name="highs", solver_options=solver_options)
+    model_status = n.optimize.solve_model(
+        solver_name="highs", solver_options=solver_options
+    )
 
     # calculate results:
-
-    def get_flh(n: Network, g: str, component_type: str) -> float:
-        if component_type == "Generator":
-            gen = n.generators_t["p"][g].mean()
-            p_nom = n.generators.at[g, "p_nom_opt"]
-        if component_type == "Link":
-            gen = n.links_t["p0"][g].mean()
-            p_nom = n.links.at[g, "p_nom_opt"]
-        if gen == 0:
-            flh = 0
-        else:
-            flh = gen / p_nom
-        return flh
 
     result_data = {}