+++ tests/test_api.py
@@ -4,6 +4,7 @@
 import logging
 import unittest
 from pathlib import Path
+from tempfile import TemporaryDirectory
 
 import numpy as np
 import pytest
@@ -26,10 +27,19 @@ class TestApi(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         """Set up code for class."""
-        cls.api = PtxboaAPI(data_dir=ptxdata_dir_static)
+        cls.temp_dir = TemporaryDirectory()
+        # create cahce dir (start context)
+        cache_dir = cls.temp_dir.__enter__()
+        cls.api = PtxboaAPI(data_dir=ptxdata_dir_static, cache_dir=cache_dir)
 
-    def _test_api_call(self, settings):
-        res, _metadata = self.api.calculate(**settings, optimize_flh=False)
+    @classmethod
+    def tearDownClass(cls):
+        """Tear down code for class."""
+        # cleanup cache dir
+        cls.temp_dir.__exit__(None, None, None)
+
+    def _test_api_call(self, settings, optimize_flh=False):
+        res, _metadata = self.api.calculate(**settings, optimize_flh=optimize_flh)
         # test that settings are in results
         for k, v in settings.items():
             if k in ["ship_own_fuel", "output_unit"]:  # skip some
@@ -92,27 +102,25 @@ class TestApi(unittest.TestCase):
         # test result categories
         res_values = res.groupby(["process_type", "cost_type"]).sum("values")["values"]
         expected_result = {
-            ("Water", "CAPEX"): 1.1342172270672746,
-            ("Water", "OPEX"): 0.539157778349926,
-            ("Water", "FLOW"): 1.195988210583831,
-            ("Electrolysis", "CAPEX"): 290.29664931801034,
-            ("Electrolysis", "OPEX"): 68.99723120650344,
-            ("Electricity generation", "CAPEX"): 927.4434737560686,
-            ("Electricity generation", "OPEX"): 308.606539951669,
-            ("Transportation (Ship)", "CAPEX"): 61.35376450396558,
-            ("Transportation (Ship)", "OPEX"): 71.15409022458118,
-            ("Transportation (Ship)", "FLOW"): 106.24743335376925,
-            ("Carbon", "CAPEX"): 196.81657679863747,
-            ("Carbon", "OPEX"): 113.57351510254443,
-            ("Carbon", "FLOW"): 100.57397108324335,
-            ("Derivative production", "CAPEX"): 90.50895668682817,
-            ("Derivative production", "OPEX"): 39.171320804334,
-            ("Heat", "FLOW"): 277.3302192660863,
-            # ("Electricity and H2 storage", "OPEX"): 176.3817106548646,
-            # TODO: test data not up to date
+            ("Carbon", "CAPEX"): 196.01309835079843,
+            ("Carbon", "OPEX"): 116.25337124331891,
+            ("Derivative production", "CAPEX"): 90.13946547212913,
+            ("Derivative production", "OPEX"): 40.0955988325783,
+            ("Electricity and H2 storage", "CAPEX"): 253.0220382601114,
+            ("Electricity and H2 storage", "OPEX"): 1.2530462794580113,
+            ("Electricity generation", "CAPEX"): 434.4326537746074,
+            ("Electricity generation", "OPEX"): 89.59100157401558,
+            ("Electrolysis", "CAPEX"): 421.1490770843375,
+            ("Electrolysis", "OPEX"): 102.17834361295175,
+            ("Heat", "FLOW"): 283.87404324169836,
+            ("Transportation (Ship)", "CAPEX"): 61.1032956243896,
+            ("Transportation (Ship)", "FLOW"): 1.823824116075305,
+            ("Transportation (Ship)", "OPEX"): 73.64164331439581,
+            ("Water", "CAPEX"): 1.1373427486958672,
+            ("Water", "OPEX"): 0.5518796288785505,
         }
 
-        result_dict = {k: pytest.approx(v) for k, v in res_values.items() if v}
+        result_dict = {k: pytest.approx(v) for k, v in sorted(res_values.items()) if v}
 
         assert expected_result == result_dict
 
@@ -134,24 +142,21 @@ class TestApi(unittest.TestCase):
         # test result categories
         res_values = res.groupby(["process_type", "cost_type"]).sum("values")["values"]
         expected_result = {
-            ("Water", "CAPEX"): 0.8511210005237106,
-            ("Water", "OPEX"): 0.21273242859187,
-            ("Water", "FLOW"): 0.471894289243873,
-            ("Electrolysis", "CAPEX"): 457.96199870023696,
-            ("Electrolysis", "OPEX"): 85.8485762832237,
-            ("Electricity generation", "CAPEX"): 838.9488831589423,
-            ("Electricity generation", "OPEX"): 146.78305821183,
-            ("Transportation (Ship)", "OPEX"): 12.2061245592688,
-            ("Carbon", "FLOW"): 61.3694736856213,
-            ("Derivative production", "CAPEX"): 188.2042202083766,
-            ("Derivative production", "OPEX"): 37.0003810701601,
-            ("Derivative production", "FLOW"): 16.8369286421993,
-            # ("Electricity and H2 storage", "OPEX"): 40.72566174247268,
-            # TODO: test data not up to date
+            ("Carbon", "FLOW"): 69.057413040518,
+            ("Derivative production", "CAPEX"): 308.1737690695886,
+            ("Derivative production", "OPEX"): 41.63553057840994,
+            ("Electricity and H2 storage", "CAPEX"): 279.89262037545114,
+            ("Electricity and H2 storage", "OPEX"): 0.5435185267752691,
+            ("Electricity generation", "CAPEX"): 842.7006740519129,
+            ("Electricity generation", "OPEX"): 92.19899158499206,
+            ("Electrolysis", "CAPEX"): 939.4974306861657,
+            ("Electrolysis", "OPEX"): 124.91778461321229,
+            ("Transportation (Ship)", "OPEX"): 13.735222655346394,
+            ("Water", "CAPEX"): 1.3502811073682484,
+            ("Water", "OPEX"): 0.23938206254847536,
         }
 
-        result_dict = {k: pytest.approx(v) for k, v in res_values.items() if v}
-
+        result_dict = {k: pytest.approx(v) for k, v in sorted(res_values.items()) if v}
         assert expected_result == result_dict
 
     def test_example_api_call_3_pipeline_sea_land(self):
@@ -171,36 +176,21 @@ class TestApi(unittest.TestCase):
         res = self._test_api_call(settings)
         # test result categories
         res_values = res.groupby(["process_type", "cost_type"]).sum("values")["values"]
-        for k, v in {
-            ("Water", "CAPEX"): 0.09778240749517693,
-            ("Water", "OPEX"): 0.0301895762734038,
-            ("Water", "FLOW"): 0.0669681098101092,
-            ("Electrolysis", "CAPEX"): 31.26888734395876,
-            ("Electrolysis", "OPEX"): 4.827015838720482,
-            ("Electrolysis", "FLOW"): 0,
-            ("Electricity generation", "CAPEX"): 222.35293638860577,
-            ("Electricity generation", "OPEX"): 58.3523144804189,
-            ("Electricity generation", "FLOW"): 0,
-            ("Transportation (Pipeline)", "CAPEX"): 5.312543588236488,
-            ("Transportation (Pipeline)", "OPEX"): 24.678148029682,
-            ("Transportation (Pipeline)", "FLOW"): 2.97636043600485,
-            ("Transportation (Ship)", "CAPEX"): 0,
-            ("Transportation (Ship)", "OPEX"): 0,
-            ("Transportation (Ship)", "FLOW"): 0,
-            ("Carbon", "CAPEX"): 0,
-            ("Carbon", "OPEX"): 0,
-            ("Carbon", "FLOW"): 0,
-            ("Derivative production", "CAPEX"): 0,
-            ("Derivative production", "OPEX"): 0,
-            ("Derivative production", "FLOW"): 0,
-            ("Heat", "CAPEX"): 0,
-            ("Heat", "OPEX"): 0,
-            ("Heat", "FLOW"): 0,
-            ("Electricity and H2 storage", "CAPEX"): 0,
-            ("Electricity and H2 storage", "OPEX"): 0,  # TODO: testdata is outdated
-            ("Electricity and H2 storage", "FLOW"): 0,
-        }.items():
-            self.assertAlmostEqual(res_values.get(k, 0), v, places=3, msg=k)
+
+        expected_result = {
+            ("Electricity and H2 storage", "CAPEX"): 44.84637867437069,
+            ("Electricity generation", "CAPEX"): 287.5830863315742,
+            ("Electricity generation", "OPEX"): 49.73468408558943,
+            ("Electrolysis", "CAPEX"): 66.31815485798717,
+            ("Electrolysis", "OPEX"): 7.180735162443889,
+            ("Transportation (Pipeline)", "CAPEX"): 8.507408889875556,
+            ("Transportation (Pipeline)", "OPEX"): 27.71899909407804,
+            ("Water", "CAPEX"): 0.15658693599049908,
+            ("Water", "OPEX"): 0.03390954768430808,
+        }
+
+        result_dict = {k: pytest.approx(v) for k, v in sorted(res_values.items()) if v}
+        assert expected_result == result_dict
 
     def test_example_api_call_4_pipeline_retrofitted(self):
         """Test output structure of api.calculate()."""
@@ -219,36 +209,18 @@ class TestApi(unittest.TestCase):
         res = self._test_api_call(settings)
         # test result categories
         res_values = res.groupby(["process_type", "cost_type"]).sum("values")["values"]
-        for k, v in {
-            ("Water", "CAPEX"): 0,
-            ("Water", "OPEX"): 0,
-            ("Water", "FLOW"): 0.362589872395495,
-            ("Electrolysis", "CAPEX"): 14.26507991048662,
-            ("Electrolysis", "OPEX"): 3.546461787521432,
-            ("Electrolysis", "FLOW"): 0,
-            ("Electricity generation", "CAPEX"): 26.50698388538755,
-            ("Electricity generation", "OPEX"): 6.58993893072181,
-            ("Electricity generation", "FLOW"): 0,
-            ("Transportation (Pipeline)", "CAPEX"): 1.281983377858589,
-            ("Transportation (Pipeline)", "OPEX"): 2.55172143395481,
-            ("Transportation (Pipeline)", "FLOW"): 0.710683430261875,
-            ("Transportation (Ship)", "CAPEX"): 0,
-            ("Transportation (Ship)", "OPEX"): 0,
-            ("Transportation (Ship)", "FLOW"): 0,
-            ("Carbon", "CAPEX"): 0,
-            ("Carbon", "OPEX"): 0,
-            ("Carbon", "FLOW"): 0,
-            ("Derivative production", "CAPEX"): 0,
-            ("Derivative production", "OPEX"): 0,
-            ("Derivative production", "FLOW"): 0,
-            ("Heat", "CAPEX"): 0,
-            ("Heat", "OPEX"): 0,
-            ("Heat", "FLOW"): 0,
-            ("Electricity and H2 storage", "CAPEX"): 0,
-            ("Electricity and H2 storage", "OPEX"): 0,  # TODO: test data is outdated
-            ("Electricity and H2 storage", "FLOW"): 0,
-        }.items():
-            self.assertAlmostEqual(res_values.get(k, 0), v, places=3, msg=k)
+        expected_result = {
+            ("Electricity and H2 storage", "CAPEX"): 18.351515364244914,
+            ("Electricity generation", "CAPEX"): 21.946684576122536,
+            ("Electricity generation", "OPEX"): 7.604789526052963,
+            ("Electrolysis", "CAPEX"): 20.875788542159256,
+            ("Electrolysis", "OPEX"): 5.384262578837443,
+            ("Transportation (Pipeline)", "CAPEX"): 1.4165302981169516,
+            ("Transportation (Pipeline)", "OPEX"): 2.925088442220974,
+            ("Water", "FLOW"): 0.41564389862359896,
+        }
+        result_dict = {k: pytest.approx(v) for k, v in sorted(res_values.items()) if v}
+        assert expected_result == result_dict
 
     def test_api_get_input_data_output_format(self):
         """Test output structure of api.get_input_data()."""
@@ -310,7 +282,7 @@ class TestApi(unittest.TestCase):
             cache_dir=None,
         )
 
-        expected_val = 0.0503
+        expected_val = 0.046
 
         # WACC without source_region_code (this is NOT fine)
         self.assertRaises(
@@ -373,6 +345,43 @@ class TestApi(unittest.TestCase):
         self.assertAlmostEqual(annuity(0.1, 10, 1), 0.162745394882512)
         self.assertAlmostEqual(annuity(0.5, 10, 1), 0.5088237828522)
         self.assertAlmostEqual(annuity(1, 10, 1), 1.0009775171)
+
+    def test_issue_553_storage_cost(self):
+        """See https://github.com/agoenergy/ptx-boa/issues/553."""
+        settings = {
+            "region": "United Arab Emirates",
+            "country": "Germany",
+            "chain": "Ammonia (AEL) + reconv. to H2",
+            "res_gen": "PV tilted",
+            "scenario": "2040 (medium)",
+            "secproc_co2": "Direct Air Capture",
+            "secproc_water": "Sea Water desalination",
+            "transport": "Ship",
+            "ship_own_fuel": False,
+            "output_unit": "USD/t",
+        }
+        res = self._test_api_call(settings, optimize_flh=False)
+        res_opt = self._test_api_call(settings, optimize_flh=True)
+        self.assertAlmostEqual(
+            res.at[("Electricity and H2 storage", "CAPEX"), "values"],
+            823.7733133374624,
+            places=4,
+        )
+        self.assertAlmostEqual(
+            res.at[("Electricity and H2 storage", "OPEX"), "values"],
+            4.079589637694555,
+            places=4,
+        )
+        self.assertAlmostEqual(
+            res_opt.at[("Electricity and H2 storage", "CAPEX"), "values"],
+            183.481609334802,
+            places=4,
+        )
+        self.assertAlmostEqual(
+            res_opt.at[("Electricity and H2 storage", "OPEX"), "values"],
+            5.567013499284561,
+            places=4,
+        )
 
 
 class TestRegression(unittest.TestCase):